requirements.txt
Flask
PyJWT
scikit-learn
redis

